adds plugin to processing step
