+++ packages/shared-db/src/index.ts
@@ -1,19 +1,14 @@
-import * as schema from "./schema";
-import * as validators from "./validators";
-
-export { schema, validators };
-
+export * as schema from "./schema";
 export * from "./schema";
-export * from "./validators";
-
+export * from "./types";
 export * from "./repositories/activity.repository";
+export * from "./repositories/authRequest.repository";
 export * from "./repositories/feed.repository";
 export * from "./repositories/feedRecap.repository";
-export * from "./repositories/twitter.repository";
 export * from "./repositories/leaderboard.repository";
-export * from "./repositories/submission.repository";
 export * from "./repositories/moderation.repository";
-export * from "./repositories/user.repository";
 export * from "./repositories/plugin.repository";
-export * from "./repositories/authRequest.repository";
 export * from "./repositories/processing.repository";
+export * from "./repositories/submission.repository";
+export * from "./repositories/twitter.repository";
+export * from "./repositories/user.repository";
