+++ packages/shared-db/src/repositories/leaderboard.repository.ts
@@ -1,7 +1,7 @@
 import { and, desc, eq, gte, sql } from "drizzle-orm";
 import * as schema from "../schema";
 import { executeWithRetry, withErrorHandling } from "../utils";
-import { DB } from "../validators";
+import type { DB } from "../types";
 
 export interface FeedSubmissionCount {
   feedId: string;
