+++ packages/shared-db/src/repositories/submission.repository.ts
@@ -1,15 +1,15 @@
 import { and, asc, count, desc, eq, ilike, or, sql, SQL } from "drizzle-orm";
 import * as schema from "../schema";
-import { SubmissionStatus } from "../schema";
-import { SelectModerationHistory } from "../schema/moderation";
-import { executeWithRetry, withErrorHandling } from "../utils";
 import {
-  DB,
   InsertSubmission,
   RichSubmission,
   SelectSubmission,
   SelectSubmissionFeed,
-} from "../validators";
+  SubmissionStatus,
+} from "../schema";
+import { SelectModerationHistory } from "../schema/moderation";
+import type { DB } from "../types";
+import { executeWithRetry, withErrorHandling } from "../utils";
 
 export interface PaginationMetadata {
   page?: number;
