+++ packages/shared-db/src/types.ts
@@ -1,0 +1,5 @@
+import { NodePgDatabase } from "drizzle-orm/node-postgres";
+
+import * as schema from "./schema";
+
+export type DB = NodePgDatabase<typeof schema>;
